Move imagery hooks to new file, addImageryProviderHooks.js
